+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
   - Restore Custom Settings values
   - Smart restore of SAML SSO Config by prompting the user to select a valid certificate
 - Send path to command log file to WebSocketServer
+- Improve startup performances by checking for sfdx-hardis upgrades every 6h and not every 15 mn!
 
 ## [6.0.6 (beta)] 2025-08-17