+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
 - Auto-detect which login type to use depending if we are in local or web context (Code Builder, CodeSpaces)
+- Add documentation for Ubuntu images
 
 ## [6.1.0] 2025-08-23