+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 
 - Auto-detect which login type to use depending if we are in local or web context (Code Builder, CodeSpaces)
 - Add documentation for Ubuntu images
+- Wait for WebSocket Server to be initialized before continuing command.
 
 ## [6.1.0] 2025-08-23